requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.36.6
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.1
constantly==15.1.0
cryptography==41.0.6
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.0
django-filter==23.3
django-rest-multiple-models==2.1.3
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
grpcio==1.59.0
grpcio-tools==1.59.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
mysqlclient==2.2.7
oauthlib==3.2.2
paramiko==3.5.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymysql==1.1.1
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.0.7
zope.interface==6.1